apps/ai-service/pyproject.toml
[tool.poetry]
name = "ai-service"
version = "0.1.0"
description = ""
authors = ["Memorysaver <mfa1484@gmail.com>"]
package-mode = false
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.11,<3.12"
uvicorn = "^0.30.6"
python-dotenv = "^1.0.1"
pillow = "^10.4.0"
requests = "^2.32.3"
websockets = "^13.1"
litellm = "^1.52.9"
instructor = "^1.6.4"
fastapi = "^0.115.5"

[virtualenvs]
# /Users/memorysaver/Documents/github/letsinfo_ai/be_ai_service/.venv
in-project = true

[tool.pyright]
# https://github.com/microsoft/pyright/blob/main/docs/configuration.md
useLibraryCodeForTypes = true
exclude = [".cache"]

[tool.ruff]
# https://beta.ruff.rs/docs/configuration/
select = ['E', 'W', 'F', 'I', 'B', 'C4', 'ARG', 'SIM']
ignore = ['W291', 'W292', 'W293']

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"